+++ src/server.ts
@@ -1,4 +1,5 @@
 import type { Writable } from "node:stream";
+import { AsyncLocalStorage } from "node:async_hooks";
 import { createElement } from "react";
 import type { FunctionComponent } from "react";
 
@@ -24,15 +25,16 @@ export type RenderInput =
       args: unknown[];
     };
 
-export type RenderOptions = {
+export type RenderOptions<Context> = {
   command: "dev" | "build" | "start";
+  ctx?: Context;
   moduleIdCallback?: (id: string) => void;
 };
 
 export type GetBuildConfig = (
   unstable_renderRSC: (
     input: RenderInput,
-    options: Omit<RenderOptions, "command">
+    options: Omit<RenderOptions<never>, "command">
   ) => Promise<PipeableStream>
 ) => Promise<{
   [pathStr: string]: {
@@ -40,6 +42,7 @@ export type GetBuildConfig = (
       readonly [rscId: string, props: unknown, skipPrefetch?: boolean]
     >;
     customCode?: string; // optional code to inject
+    ctx?: unknown;
     skipSsr?: boolean;
   };
 }>;
@@ -64,4 +67,28 @@ export function ClientFallback() {
 // For internal use only
 export function ClientOnly() {
   throw new Error("Client-only component");
+}
+
+const ContextStore = new AsyncLocalStorage();
+// FIXME this is not what we want
+(globalThis as any).WAKU_SERVER_CONTEXT_STORE ||= ContextStore;
+
+export function getContext<T>() {
+  const ContextStore: AsyncLocalStorage<unknown> = (globalThis as any)
+    .WAKU_SERVER_CONTEXT_STORE;
+  const ctx = ContextStore.getStore();
+  if (ctx === undefined) {
+    throw new Error("Missing runWithContext");
+  }
+  return ctx as T;
+}
+
+// For internal use only
+export function runWithContext<Context, Result>(
+  ctx: Context,
+  fn: () => Result
+): Result {
+  const ContextStore: AsyncLocalStorage<unknown> = (globalThis as any)
+    .WAKU_SERVER_CONTEXT_STORE;
+  return ContextStore.run(ctx, fn);
 }